# api-sync-gen/examples/nestjs-sample-app/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TourMate API",
    "description": "Travel tour booking platform API",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "Auth"
        ],
        "security": []
      }
    },
    "/bookings": {
      "get": {
        "operationId": "BookingsController_findAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "BookingsController_create",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tourId": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "startDate": {
                    "type": "string"
                  },
                  "numberOfPeople": {
                    "type": "number",
                    "minimum": 1
                  }
                },
                "required": [
                  "tourId",
                  "userId",
                  "startDate",
                  "numberOfPeople"
                ]
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}": {
      "get": {
        "operationId": "BookingsController_findOne",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "BookingsController_remove",
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}/cancel": {
      "patch": {
        "operationId": "BookingsController_cancel",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/tours": {
      "get": {
        "operationId": "ToursController_findAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tours"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": []
      },
      "post": {
        "operationId": "ToursController_create",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Tours"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 2000
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 999999
                  },
                  "duration": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 365
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "adventure",
                      "cultural",
                      "wildlife",
                      "beach",
                      "mountain"
                    ]
                  },
                  "difficulty": {
                    "type": "string",
                    "enum": [
                      "easy",
                      "moderate",
                      "hard"
                    ]
                  },
                  "maxGroupSize": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 50
                  },
                  "highlights": {
                    "type": "string"
                  },
                  "startLocation": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "description",
                  "price",
                  "duration",
                  "category",
                  "difficulty",
                  "maxGroupSize"
                ]
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/tours/featured": {
      "get": {
        "operationId": "ToursController_getFeatured",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tours"
        ],
        "security": []
      }
    },
    "/tours/{id}": {
      "get": {
        "operationId": "ToursController_findOne",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Tours"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": []
      },
      "patch": {
        "operationId": "ToursController_update",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Tours"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 100
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0
                  },
                  "category": {
                    "type": "string"
                  },
                  "difficulty": {
                    "type": "string"
                  },
                  "maxGroupSize": {
                    "type": "number",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "ToursController_remove",
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "Tours"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/tours/{id}/bookings": {
      "get": {
        "operationId": "ToursController_getTourBookings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tours"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": []
      }
    },
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "UsersController_register",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50
                  },
                  "lastName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "phone": {
                    "type": "string"
                  }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "security": []
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "tourId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "numberOfPeople": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "tourId",
          "userId",
          "startDate",
          "numberOfPeople"
        ]
      },
      "CreateTourDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "minLength": 20,
            "maxLength": 2000
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "maximum": 999999
          },
          "duration": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          },
          "category": {
            "type": "string",
            "enum": [
              "adventure",
              "cultural",
              "wildlife",
              "beach",
              "mountain"
            ]
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "moderate",
              "hard"
            ]
          },
          "maxGroupSize": {
            "type": "number",
            "minimum": 1,
            "maximum": 50
          },
          "highlights": {
            "type": "string"
          },
          "startLocation": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "price",
          "duration",
          "category",
          "difficulty",
          "maxGroupSize"
        ]
      },
      "UpdateTourDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "category": {
            "type": "string"
          },
          "difficulty": {
            "type": "string"
          },
          "maxGroupSize": {
            "type": "number",
            "minimum": 1
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "lastName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://staging-api.tourmate.com",
      "description": "Staging"
    }
  ]
}